refactor progress screen mock data and utilities
